Order the files in the settings directory, move files directly related only to email theme editing into here.
