+++ src/features/settings/layout/EmailThemeLayout.tsx
@@ -1,10 +1,9 @@
 import { FunctionComponent } from 'react';
 
-import messageIds from 'features/emails/l10n/messageIds';
+import messageIds from 'features/settings/l10n/messageIds';
 import { useMessages } from 'core/i18n';
 import { useNumericRouteParams } from 'core/hooks';
 import SimpleLayout from 'utils/layout/SimpleLayout';
-import ThemeLayoutActionButtons from 'features/emails/components/ThemeLayoutActionButtons';
 
 interface EmailThemeLayoutProps {
   children: React.ReactNode;
@@ -16,15 +15,11 @@ const EmailThemeLayout: FunctionComponent<EmailThemeLayoutProps> = ({
   const { themeId } = useNumericRouteParams();
   const messages = useMessages(messageIds);
   const title = themeId
-    ? messages.themes.themeEditor.title({ themeId: themeId })
-    : messages.themes.title();
+    ? messages.email.themes.themeEditor.title({ themeId: themeId })
+    : messages.email.themes.title();
 
   return (
-    <SimpleLayout
-      actionButtons={<ThemeLayoutActionButtons />}
-      fixedHeight={!!themeId}
-      title={title}
-    >
+    <SimpleLayout fixedHeight={true} title={title}>
       {children}
     </SimpleLayout>
   );
